refine(core): 修正数据模型类型、注入时间校准基准并优化CoreDataBus API
1. 统一Voltage_block.SampleCount类型为int，移除冗余强制转换
2. CoreBusHeader新增双时间校准字段，消除跨进程时钟歧义
3. ReadLatestSingle重构为TryReadLatestSingle，明确数据返回语义
4. 下调CoreDataBus缓冲区容量至100万采样点，降低内存占用
5. 新增CONTEXT.md统一领域术语定义

+++ ConsoleApp1/Models/Voltage_block.cs
@@ -17,7 +17,7 @@ namespace AvaloniaApplication1.Models
         
         // 如果同步双通道采样，则每个通道采样点=nBytes/4（每个采样点2字节）
         // 如果单通道采样，则每个通道采样点=nBytes/2（每个采样点2字节）
-        public double SampleCount;   // 每个通道的采样点数量
+        public int SampleCount;   // 每个通道的采样点数量
         public long StartTick;       // 起始时间戳（高精度）
     }
 }
